+++ R/AllGenerics.R
@@ -460,31 +460,3 @@ setGeneric(
   name = "jackknife",
   def = function(object, ...) standardGeneric("jackknife")
 )
-
-# Deprecated ===================================================================
-#' Deprecated Methods
-#'
-#' @param from An object to be coerced.
-#' @param factor A [`logical`] scalar: should character string be
-#'  coerced to [`factor`]? Default to `FALSE`, if `TRUE` the original ordering is
-#'  preserved.
-#' @param reverse A [`logical`] scalar: should the order of factor levels be
-#'  reversed? Only used if `factor` is `TRUE`. Useful for plotting.
-#' @param object A [`numeric`] vector.
-#' @param level A length-one [`numeric`] vector giving the confidence level.
-#'  Must be a single number between \eqn{0} and \eqn{1}.
-#' @param type A [`character`] string giving the type of confidence
-#'  interval to be returned. It must be one "`student`" (the default) or
-#'  "`normal`". Any unambiguous substring can be given.
-#' @param x,y A [`numeric`] object to be checked.
-#' @param expected A [`character`] string specifying the expected
-#'  value (see details).
-#' @param margin A vector giving the subscripts which the function will be
-#'  applied over (`1` indicates rows, `2` indicates columns).
-#' @param ... Extra parameters to be passed to internal methods.
-#' @author N. Frerebeau
-#' @docType methods
-#' @name deprecate
-#' @rdname deprecate
-#' @keywords internal
-NULL